Consolidate form validation schemas into their own directory `./src/forms` (#2570)
