+++ .eslintrc
@@ -116,6 +116,7 @@
           ["^(store)(/.*|$)"],
           ["^(types)(/.*|$)"],
           ["^(utils)(/.*|$)"],
+          ["^(forms)(/.*|$)"],
           ["^(hooks)(/.*|$)"],
           ["^(pages)(/.*|$)"],
           ["^(components/templates)(/.*|$)"],
