+++ src/forms/createUrlIfNoFileSchema.ts
@@ -1,0 +1,10 @@
+import * as Yup from "yup";
+
+export const createUrlIfNoFileSchema = (fieldName: string) =>
+  Yup.string().when(
+    fieldName,
+    (file: FileList | undefined, schema: Yup.MixedSchema<FileList>) =>
+      file && file.length > 0
+        ? schema.notRequired()
+        : schema.required("Required")
+  );
