+++ src/forms/worldEntranceSchema.tsx
@@ -1,0 +1,12 @@
+import * as Yup from "yup";
+
+import { entranceSchema } from "forms/entranceSchema";
+import { questionSchema } from "forms/questionSchema";
+
+export const worldEntranceSchema = Yup.object().shape({
+  adultContent: Yup.bool().notRequired(),
+  requiresDateOfBirth: Yup.bool().notRequired(),
+  code: questionSchema.notRequired(),
+  profile: questionSchema.notRequired(),
+  entrance: entranceSchema.notRequired(),
+});
